Rhizome changes for storage device support
Adds `volume["storage_device"]` to storage volume parameters hash map.
This effects where the files of a volume is going to be stored.

If `volume["storage_device"]` is `"DEFAULT"`, then we do the same thing
as before and we put it under `/var/storage/#{vm_name}/#{disk_index}`.

If it is something else, then it is stored under
`/var/storage/devices/#{device_name}/#{vm_name}/#{disk_index}`.

The above path logic is put into `storage_path.rb`, and other places
that need to know about the file structure uses that.