Add default dns server to dnsmasq.conf
We need to add the default dns server setting to dnsmasq.conf to make
alma linux name resolution work. Looks like the DHCP packets from
dnsmasq contains a DNS field which is filled with the VM facing
interface's ip address when the option is not specified. That is not
picked up by the linux VMs but almalinux does. That results in name
resolution failures which causes overall network failure. Here, we set
it to quad9 servers which fixes the issue.