Add extra logging when leases are unexpectedly cleared
Unfortunately, this incurs an extra query returning a lot more data
(relatively) and transaction block on every freeing of a lease, but
it's a cost we can incur.

After increasing parallelism levels, we have still yet another bug
rattling around with transient strands, probably still related to
`exitval`.  It's rare: for the many, many strand executions, I've seen
it once since I made that change.

Adding this information can help us with vital clues, such as the
state of the `lease` and `exitval` columns.