Fix the bug that waiting workflow jobs steal the runners
Today, I noticed a GithubRunner instance running for ~8 hours. This
seemed suspicious to me, as runners are typically ephemeral and CI jobs
generally do not last this long. I suspected that the runner might be
stuck. I checked and the job status was "waiting". GitHub Actions has a
manual approval mechanism. The job remains in the "waiting" state until
someone approves it. The job states proceed in the following order:
waiting, queued, in_progress, and completed. If a job doesn't require
manual approval, it begins in the "queued" state. We provision a new
runner when we receive a webhook event for a workflow job in the
"queued" state.

GitHubRunner rows have an empty runner_id column until the runner picks
up a job. When GitHub sends a workflow_job event in the "waiting" state,
`workflow_job["runner_id"]` is also empty. Therefore, even when the job
is in the "waiting" state, we assume that the runner_id values match and
associate the job with the runner.  Consequently, the runner waits for
the job, which remains in the "waiting" state until manual approval.