Add deadline for DNS purge
We observed a case where DNS purge stuck because some of the rows it tries to
delete were already deleted, which is unexpected because the only place that
deletes DNS records is purge_dns_records label. While debugging the issue we
fixed it thus missed the furter debugging chance to understand root cause. I
couldn't reproduce this in my local environment, so I'm adding a deadline in
case we hit it again.