Redirect the invited users to the project's dashboard
When a user is invited to a project, the admin also needs to grant them
permission. The only page accessible without permission is the project's
dashboard page. However, we currently redirect invited users to the
project settings page, which results in a 403 error. It's a better user
experience to redirect them to the dashboard page.