Update dnsmasq to v2.89
I previously used a semi-arbitrary commit to pick up some recent IPv6
related changes.  I am not sure it was relevant: it was a long time
ago, with the service in a more primitive state.

But it's a reasonable time to use a dnsmasq release, now that one is
newer than the hash I previously recorded.

You can verify this at hash and tag at:

https://thekelleys.org.uk/gitweb/?p=dnsmasq.git;a=commit;h=5dc14b6e05f39a5ab0dc02e376b1d7da2fda5bc1