Accept Hetzner API 404 for delete_key as success
If it is gone, there's no need to halt execution.  This can cause CI
runs to fail, in the case where a key was deleted but, for one reason
or another, the deletion success was not recorded.

Perhaps our API should not inflict a similar damage on the consumer,
see
https://github.com/ubicloud/ubicloud/issues/597#issuecomment-1718348587