MinIO Scale-out support by adding new pool
With this commit, we are adding support to add new pools to an existing
MinIO cluster. The documentation is here
https://min.io/docs/minio/linux/operations/install-deploy-manage/expand-minio-deployment.html
The idea is to create a new pool with servers, reconfigure the cluster
by editing /etc/default/minio in all servers and perform a restart.
Servers detect the new pool automatically and perform the metadata
exchange.