Provision a new runner when possible missing runner
Nowadays, GitHub doesn't send some "workflow_job.queue" webhook events
to us. We provision fewer runners than necessary because of missing
webhook deliveries.
If GitHub cannot reliably deliver all webhooks to us, the ultimate
solution would be to regularly poll GitHub data and detect any missing
runners. I started to work on implementing this polling system, but it
presents its own set of challenges and requires time to design.

In the interim, I have implemented a partial solution to mitigate the
effects. While this does not resolve the issue of missing webhooks, it
lessens the impact of the missing runner symptoms.

GitHub sends us a "workflow_job.in_progress" event when a runner picks
up a job. If there is a significant time difference between the job's
creation date and the runner's creation date, it's possible that an
older job has claimed a new runner. In such cases, we register a new
runner for that repository to compensate for the missing one. If this
turns out to be a false positive, our runners will automatically
deprovision themselves if they don't pick up a job within two minutes.
Additionally, we do not charge customers for runners that do not pick up
a job.
Once we have implemented the polling feature, we can rollback this
feature. After implementing polling, we should no longer experience
issues with missing runners.