Make interface setup idempotent
We found this thanks to a page. A vm stuck while setting up interfaces. Upon
close inspection, we found that, in the `interfaces` function, we catch some
exceptions with various error messages, but we don't resume the execution in
that function. We need to catch exceptions for each command separately and
resume the execution with the next command in the function. However there is
a command that generates its inputs (MAC addresses randomly, which makes it
unsuitable for error message matching. Instead, just deleting and recreating
the network namespace is easier and safer alternative to achieve idempotency,
so we went with that.