Single instance MinIO configuration fix
The single instance MinIO clusters do not get the host address prefix
before the mount point because there is no need to connect to any other
node. Therefore, the configuration must simply give a local mount point
path. With this commit, we fix that, before this, the provisioning was
failing for single instance configuration.