Refactor MinIO Provisioning Path
With this commit we are changing the provisioning path for MinIO
cluster. The main objective is to eliminate the redundant columns from
the MinIO cluster. They were necessary since the pool or server
provisioning was happening in later stages of the state machine. Moving
all of the entity creations into the assemble functions, we eliminate
the need of redundant columns and load those values dynamically.