Run CI on both x64 and arm64 runners
The code for our rhizome (dataplane) depends on the architecture of the
VM host. It's possible for tests to pass on x64 but fail on arm64 due to
incorrect assumptions, such as f58dc74a. I believe it's beneficial to
run rhizome tests on both architectures.

Additionally, testing our arm64 runners is a sound strategy.

The "ruby/ruby-builder" currently lacks an arm64 release for Ubuntu
[^1], preventing us from adding Ruby to our golden ARM image using the
default scripts [^2]. I prefer not to create custom scripts for our
golden image.  However, if multiple customers request Ruby on arm64
images, I'll consider it.  At present, I only install and cache it for
arm64 runners.

[^1]: https://github.com/ruby/setup-ruby#supported-platforms
[^2]: https://github.com/actions/runner-images/blob/main/images/ubuntu/scripts/build/install-ruby.sh