Apply new Rubocop corrections
Dependabot sent a pull request for rubocop-related updates, and it
failed because of new corrections that could be applied.  This applies
them, and then Dependabot's rebase of the rubocop update should
succeed.

The default value for `Capybara/ClickLinkOrButtonStyle` has been changed
from `strict` to `link_or_button`, which has loosened its constraints.
However, we have been adhering to the `strict` style. I've reverted it
back to `strict` as the relaxed version necessitated numerous test
changes, which I found to be unnecessary.

https://github.com/rubocop/rubocop-capybara/pull/76