Use 2 vcpus for SPDK on new hosts.
Each OS vcpu can only use one NVMe queue, and in our benchmarks we can't
saturate a NVMe card using a single queue. This patch changes number of
vcpus used for SPDK to 2 to improve the overall I/O performance.

Note that a single VM will be limited to the performance of a single
NVMe queue, as SPDK's vhost_user_blk implementation uses a single
thread. But running multiple VMs in a host will allow using both vcpus.

In an AX-161 machine:
* 1-vcpu, 1 or more VMs: unencrypted 697 MiB/s, encrypted 540 MiB/s
* 2-vcpu, 2 or more VMs: unencrypted 1360 MiB/s, encrypted 1040 MiB/s

We could get better performance in encrypted case by using 3 vcpus, but
then we would decrease VM capacity in a host by 1. Assuming 2-core VMs,
in a host with 32 total cores, assigning 1 or 2 vcpus for SPDK allows 31
VMs, but 3 vcpus for SPDK allows 30 VMs. So we're accepting this
trade-off for now.