Add monitor process
We are adding a new executable for health monitor. This process would fetch the
list of resources that needs to be monitored and checks their health in a loop.

Here are the things we specifically cared about while implementing this;
- We are decoupling health checks from the resource's nexus progs. This allows
us to perform very frequent health checks independent from the nexus' schedule.
It also does not cause unnecessary lock holding while doing the health checks.
Finally this allows us to independently scale monitoring process.
- We use persistent SSH sessions for monitoring. This ensures very fast health
check cycles. Moreover, we even use SSH tunnels to communicate with the target
software directly, thus avoiding many expensive system calls and forks, though
example of such communication is not available in this commit. See the next one
for an example.
- Each resource decides how the persistent session must be created and how the
check itself must be performed. It does this by defining two functions called;
init_health_monitor_session and check_pulse. Monitoring process creates the
session using init_health_monitor_session and passes it to check_pulse. Primary
contract between the monitor and the resource is that the session needs to be
based on SSH. In a similar fashion, resource can decide how frequently it needs
to be monitored by defining monitoring_interval function.