Add helper to sshable to start fresh SSH sessions
In the context of monitoring, we want a fresh SSH session that is not taken
from the thread's cache, because we are calling session.process ourselves and
doing that interferes with regular command execution. To prevent that, we are
creating a fresh SSH session for the resources we want to monitor.