Try to recover Postgres from unavailability
Currently we just try to restart PostgreSQL process. In the future, this can be
extended to triggering failover for HA enabled databases or creating a standby
for non-HA databases.

One usual false positive is database not responding to queries during the crash
recovery. In such cases, triggering restart is counter-productive as it would
undo some part of crash recovery. To prevent that, we check the logs and not
trigger restart if the database is in crash recovery.