Don't schedule disks on bdev_ubi enabled hosts if not using bdev_ubi.
Currently, github runners require btrfs if not using bdev_ubi. We will
use ext4 for hosts with bdev_ubi installation. So, we make sure we don't
schedule disks on a bdev_ubi enabled host if bdev_ubi is not used.

Note that this is a temporary measure, and can be removed when either we
have a direct way to check for btrfs (e.g. if we add this field for
storage devices), or when we migrate away from btrfs completely.