Simplify Vm allocate tests
Per usual reflex, these tests started out attempting to avoid saving
records to the database by default, such that tests that can avoid it
can remain efficient.

But as the allocation code is mostly SQL, every test in the describe
block, save one, opted into creating real records to get a useful
test.

Concede that this is the right default for this describe block, and
make it the default behavior, reducing the number of expressions.

Co-authored-by: Hadi Moshayedi <hadi@ubicloud.com>