Introduce new billing rates for arm64 runners
We decided to keep the same pricing for arm64 runners as x64 runners. We
don't require new billing rates for them. However, when we use the same
billing rate, we can't differentiate between x64/arm64 runner usage in
reports and invoices.

It would be more efficient to use separate billing rates for arm64
runners, even if the unit price remains the same.