Add storage flags to vm pools.
Previously it was possible that we return a VM with wrong storage flags
from a vm pool, because vm pools were not filtered by storage flags.

This PR adds storage flags to vm pools to fix that.

Note that we only include flags that affect the storage device
guarantees. These flags are:
* `skip_sync`: which ignores FLUSH requests to the block device. This
  increases the performance, but makes the device not crash-recoverable.
* `encrypted`

We didn't include `use_bdev_ubi` because it only affects performance,
and not the storage guarantees.