Migrations for core accounting.
== Schema changes ==

We made couple of recent observations:
- We need to use more than 1 core for SPDK to saturate NVMe cards
- Allocating 2 SPDK services on same cores reduces performance.

For this reason, we are adding core_count and core_offset fields to
SpdkInstallation. We will use these fields for SPDK service provisioning
in a future PR. In this PR, we are just using them to take into account
SPDK cores when doing vm allocation.

Note that having 2 SPDK installations on the same host (which reduces a
host's capacity) will only happen when doing migration and will be a
temporary situation.

Also, we are dropping the core_allocation_limit, (1) we don't want to do
have core reservation logic scattered in both code & db constraints, (2)
we won't use VmHost.used_cores anymore.

We will drop use_cores in a future PR.

== vm_host_with_stats view ==

This is a new view which has live unmaterialized values of vm count &
cores + spdk count & cores. This view will be used in the next commit to
get live used cores instead of using a materialized used_cores.