Query used cores directly from provisioned Vm and Spdks.
Previously we updated VmHost::used_cores to keep track of used cores.
Keeping this value adds complexity to the code, e.g. we need to think
about concurrency when updating it, and we might forget to update it in
some situations when making changes (as we did when installing spdk).

This PR switches to a new method which gets this value directly from Vm
and SpdkInstallation records created on the host.

This method can also be applied to used hugepages, which we can do in a
follow-up PR.

Also we also need to update SPDK core assignment in a follow-up PR.