Freeze core classes in production redux
I could produce the faults in 5b6cf623a858615487c27af7e306f2011d1467f0
with:

    RACK_ENV=production foreman start

That's what I thought I did before to test this patch, but maybe I
typoed something.

Analyzing them, I saw that there were two problems:

1. The eager loading code for production in loader.rb is incomplete;
   some specially handled symbols need to be in the
   `AUTOLOAD_CONSTANTS` array.

2. Various entry-points (as in being in `bin` or in `config.ru`) had
   varying set-up requirements before freezing.

The second problem is tricky.  The option I elected for here is to
require every entry point (so everything in `bin` plus something that
hooks into `puma`'s invocation) to be cognizent of when to freeze.  In
at least one case, `bin/pry`, I elected not to freeze.

An alternative, which perhaps is feasible but needs more research,
would be to insist that `bin` entry points do not make any
modifications to Object, forcing them to push most of their logic into
autoloaded symbols.  That is a plausible programming rule but should
be tried before we commit to it.  Since I don't have time for that,
I'm opting for the more invasive approach to keep the code more
undisturbed.