Freeze core classes in production
Alas, we can't really run this through unit tests, since we use
`rspec-mock`, which mutates classes in order to test things.  As a
small consolidation, it cleans up the branches seen in
roda-sequel-stack.

But even if we could test with freezing of classes...library gems are
entitled to do weird core-class modifications that can break.  To some
extent, integration testing is the only way for this one.