Download boot image from MinIO cluster with presigned url
We decided to distribute our custom image through our internal MinIO
cluster instead of Azure Blob Storage, with this commit marking the
latest step in that process.

Presigned URLs are a convenient solution as they eliminate the need to
send MinIO credentials to VM hosts. They expire after a specified
period, enhancing their security for dataplane transmission.

If a custom URL isn't provided to the "download_boot_image" prog, it
generates a presigned URL and uses it to download the image.

This approach also simplifies the process for developers to download
images for GitHub runners. They only need to set the relevant
configurations and initiate the download with the command
`vm_host.download_boot_image("github-ubuntu-2204")`.

Initially, I considered using the `mc` client to download images from
the MinIO cluster. However, it seems that the `mc` client only supports
downloads using the access key and secret key, and does not support
downloads with presigned URLs. I attempted to use `curl` for the
download, which performed similarly to the `mc` client. Therefore, I
opted for the simpler option and decided to proceed with `curl` for now.