Add rhizome multi-device storage abstractions
Managing multiple storage devices per VM indicates makes the old
one-to-one assumptions of the code in `vm_path.rb` obsolete.  This
patch introduces a similar `storage_path.rb` that is capable of
computing paths for multiple storage devices.

With the exception of the storage-key-tool interface change -- largely
harmless as the key tool is not run automatically -- these rhizome
changes are thought to be backwards compatible, and can be deployed
first.

Hadi wrote the code, but I am taking responsibility for breaking it up
and deploying it.

Co-authored-by: Hadi Moshayedi <hadi@ubicloud.com>