Fix double delete in DNS purge
If the record we want to delete exists in both obsoleted_records list and
seen_tombstoned_records list, we try to delete them twice. Second deletion
fails with the error message "Attempt to delete object did not result in a
single row modification". To prevent double deletion, we deduplicate the
records.