Run E2E tests with respirate
We dream of running E2E tests with the respirate. Currently, the
`bin/ci` script assembles a `Test::HetznerServer` strand and runs it
until termination. This is a primitive version of respirate, assuming
all other strands are its children. For GitHub E2E tests, we also need
to provision `GithubRunner` strands via a webhook, which are not
children of `Test::HetznerServer`. Alongside this, we need to run a web
server. I used `foreman` to run multiple services, similar to Heroku's
approach. It exists when one of the services exit.

Running tests with the respirate has its advantages and disadvantages.
However, I believe the advantages outweigh the disadvantages.

**Pros:**

- We also test the respirate functionality in E2E tests.
- We don't need to develop custom functionality like parent-child
  relationships on all strands. This doesn't align with our other progs,
  and it's unnecessary for the respirate. In the next commit, I removed
  over 200 lines.
- The primitive runs one strand at a time, while the respirate executes
  them concurrently.
- While the primitive waits on naps, the respirate runs other strands.
- The primitive has a custom retry mechanism, but the respirate uses its
  own retry mechanism.

**Cons:**

- The primitive runs strands itself, having direct access to the results
  of the strand runs. Conversely, the respirate runs strands, which
  might get stuck at some point, causing the CI script to wait
  unnecessarily.
- The primitive provides superior logging. However, the new version
  includes the respirate logs, and I have added some logs to the new CI
  script.