VM firewall rules view
This commit implements the UI components of the firewalls for VMs. We
keep it minimal and not make it a part of the provisioning path. The VM
comes with 2 default firewall rules anyway, we simply visualize them in
the detailed VM view and allow customer to add new firewall rules or
delete existing ones.