Do not auto-refresh if VM is already running
We keep auto refreshing the VM page even after the VM is provisioned and
in running state. This was initially introduced to give an update about
the statues of provisitioning at the start. However, keeping it
refreshing even after the provisioning is finished is creating a funny
situation. When we try to write a cidr value to create a new firewall
rule, if we're not fast enough, the page refreshes and the firewall rule
definition that we wanted to create gets lost. This commit simply checks
the VM[:state] and if it is running, we do not refresh.