GithubRunner and VMPool firewall refactoring
With this commit, we are removing the individual firewall rule
modifications in the GithubRunner and VMPool progs to start relying on
allow_only_ssh property of the Vm::Nexus.assemble_with_sshable. This
way, the default allow all rules that we add to VMs are not added from
the start and we simply don't need to delete them.