Wait for completion of restart during MinioServer unavailability
Previous way of canceling the restart was hazardous. Because, after we start
the restart process, there is no guarantee that it can be cancelled safely.
Even if we destroy the child strand, restart would continue as it is started
via daemonizer. It is more predictable to wait for restart's completion and
then switching to wait state. This approach does not increase unavailability,
in fact it could decrease slightly, but it can increase the duration that we
consider ourselves unavailable.