Support multiple PostgresServers per PostgresResource
So far, we assumed PostgresResource would only have one PostgresServer entity
and we wrote bunch of code with that assumption. However, in the case of high
availability and SKU changes with failovers, a PostgresResource can have more
than one PostgresServer entities. In such cases, it is important to have a way
distinguish the PostgresServer that is the primary contact point of the given
PostgresResource in places like taking backups or configuring the DNS records.
This commits introduced representative_server association for PostgresResource,
which refers to that primary contact point.

representative_server is different than the primary server. Restoring or read
replica PostgresResources would not have a primary server but they would have
a representative_server.

Rest of the commit is some helpers and boring refactors to ensure that we don't
assume only one PostgresServer per PostgresResource.