Change the location for temporary image downloads
We initially download the boot images to a temporary location before
relocating them to their final destination. The current temporary
location is "/tmp/", while the final location is "/var/storage/images".
Our GitHub runner VM hosts mount the "/var/storage" directory as a btrfs
partition. Downloading a GitHub image requires moving 86GB file between
devices, an action we aim to avoid. Furthermore, when the directories
are on the same partition, we can use "File.rename" instead of "mv",
which achieves the task with fewer system calls.