Add bdev_ubi enabled VMs to VmPool
In order to test bdev_ubi enabled VMs we were using project
specific storage policies to create VMs on the host with ubi SPDK
version. This allowed us to test bdev_ubi only for the ubicloud
project. We are done with the test and want to allow other projects
to use the VmHost with bdev_ubi.

With this commit we start to create VMs as part of VmPools with
bdev_ubi. Instead of passing the bdev_ubi information as project
specific policy, VMs hitting a VmHost with ubi SPDK version will
create VM with bdev_ubi.