Finalize the image download with File.rename
At present, we download the image to a temporary path and then move it
to the final location using the `qemu-img convert` command. However,
this process is not atomic and can be time-consuming.  It would be more
efficient to use the `File.rename` command to move the image directly to
the final location, as this is an atomic operation.  This method also
allows other virtual machines to use the old image right up until the
last moment.