Allow only one concurrent image download with flock
Right now, we use the "File::CREAT | File::EXCL" flags to open a
temporary file for image downloads, which should stop multiple downloads
happening at once. These flags only let one process create the file, and
if the file is already there, it fails. However, if a download gets
interrupted, the file sticks around and the next download attempt fails.
It's tricky to figure out if the file is incomplete, so we have to
delete it manually. Using flock to lock a lock file is a more effective
way to stop multiple downloads idempotently at the same time.