Only perform wait_sshable check on IPv4
As-is, our hosting environment cannot handle IPv6 transit and IPv6
unistacked VMs *never* reach the provisioned state successfully...even
though they are in fact available.

However, someone reading the console doesn't know this: they'd
conclude the system is totally broken.

I thought about removing the `wait_sshable` state entirely, as in the
general case people can set firewalls restrictive enough to not listen
to SSH or boot custom images that may not use SSH or configure it
differently.  This was my objection even when it was introduced.

However... as a pragmatic measure, `wait_sshable` still teaches us
about occasional boot problems, and as we do not yet support custom
images, we can be reasonably assured that failure to open this port is
abnormal.  Getting paged for those provides lessons we still need.

To mostly retain that source of information, I trim our sails to only
do this check when the targer computer has a route via IPv4.