Allow to link back to another label for continuation of push
Pushing is becoming popular these days. If you're budding a single prog
and don't require concurrency, push is a better option than bud.

At present, its usage pattern is complex. You need to check the retval
of push in the same label you pushed, then move to the next label.

However, we can simplify this by allowing a link back to another label
for the continuation of push. This way, the pushed program returns the
next label instead of the previous one.