Change user hyper tag with UBID
All resources are represented in our authorization system, referred to
as a "hyper tag". Each resource has a unique format, for example, users
are represented as "user/test@example.com", projects as
"project/pjbw1tcvkew67qcxqsjpbvh19c", and virtual machines as
"project/pjcm807azyxe6ht7ct027tcg83/location/hetzner-fsn1/vm/my-vm".

We prefer using user-friendly names over UBIDs, making it easier for
customers to edit their access policies. However, we identified a
problem with this approach. When a resource's hyper tag changes (e.g., a
user changes their email), the existing access policies need to be
updated accordingly. Without such updates, permissions to resources
could be affected.

Automatically updating access policies to apply these changes is not
recommended. Instead, I propose keeping the resource's UBID as the
hyper tag in the database, which eliminates the need for changes when
resources are renamed. This is because UBIDs remain constant throughout
a resource's lifetime.

We only need to map the prettified policy name when displaying or saving
access policies. The authorization system uses UBIDs in the background.

I have refactored the user's hyper tag for now. I will continue to
refactor other resources in subsequent PRs after the deployment of user
change.

If the user is missing, it raises an error.

We need to update user hyper tags in production after deployment.

From a security perspective, we only transform tags/UBIDs for resources
associated with this project.